memory-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
qdrant-client==1.9.1
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.55.3
httpx==0.27.2
